Move key-text extraction to patient profile fragment
